% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aae.R
\docType{data}
\name{aae}
\alias{aae}
\title{AAE values in Blantyre, Malawi}
\format{
A dataframe with 4 variables and 341 observations

\describe{
\item{time-malawi}{Local time in Malawi}
\item{bc-blue}{Black carbon concentration at infra-red wavelength light \link{ng/m3}}
\item{bc-ir}{Black carbon concentration at blue wavelength light \link{ng/m3}}
\item{aae-blue-ir}{Absorption angstrom exponent at Blue/IR wavelength}
}
}
\usage{
aae
}
\description{
The dataset consist of Black carbon concentrations in the breathing zone in Blantyre, Malawi. The Absorption angstrom exponent
is calculated using the Black carbon concentrations at the two wavelengths.
}
\keyword{datasets}
